q2_coding_agent_recommend/backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
requests
python-dotenv
google-generativeai
pydantic
